+++ test/unit_test/task/filebased.py
@@ -1,8 +1,7 @@
 from appintegration.task.framework import ApplicationIntegrationSourceTask, ApplicationIntegrationProcessorTask
 from appintegration.task.filebased import CSVTask, XLSXTask, JSONTask, XMLTask, PropertiesTask
 
-from .._utils.file.format import FormatTestSpec
-from .._config import (
+from ..._config import (
     # For file paths
     Test_CSV_File_Path,
     Test_XLSX_File_Path,
@@ -13,6 +12,7 @@ from .._config import (
     Test_Reading_Mode, Test_Writing_Mode, Test_XML_Writing_Mode
 )
 from ..._data import Test_Data_List, Test_JSON_Data
+from .._utils.file.format import FormatTestSpec
 from ._spec import AppIntegrationTaskTestSpec, _AppIntegrationTaskGeneric
 
 from typing import Union, Iterable, Any, Generic
